// types/app.types.ts
export interface User {
  id: string;
  username: string;
  fullName: string;
  email: string;
  profileImage: string | ImageSourcePropType;
  bio?: string;
  followers: number;
  following: number;
  likes: number;
  isFollowing?: boolean;
  isVerified?: boolean;
  isOnline?: boolean;
  createdAt?: string;
}
// ...
export interface Comment {
  id: string;
  videoId?: string;
  user: User;
  text: string;
  likes: number;
  isLiked?: boolean;
  replies?: Comment[];
  createdAt: string;
}